+++ app/strategy/payoff.py
@@ -10,7 +10,7 @@ def intrinsic(price: float, strike: float, option_type: str) -> float:
 
 
 def leg_pnl_usd(leg: dict[str, Any], expiry_price: float, spot: float) -> float:
-    qty = int(leg.get("quantity") or 1)
+    qty = float(leg.get("quantity") or 1)
     premium = float(leg.get("price_btc") or 0) * spot * qty
     value = intrinsic(expiry_price, float(leg["strike"]), leg["option_type"]) * qty
     if leg["side"] == "buy":
@@ -56,7 +56,7 @@ def build_payoff(legs: list[dict[str, Any]], spot: float, target_price: float | 
 def premium_summary(legs: list[dict[str, Any]], spot: float) -> dict[str, Any]:
     net = 0.0
     for item in legs:
-        value = float(item.get("price_btc") or 0) * spot * int(item.get("quantity") or 1)
+        value = float(item.get("price_btc") or 0) * spot * float(item.get("quantity") or 1)
         net += -value if item["side"] == "buy" else value
     return {
         "net_premium_usd": net,
